Break operation customization codegen into its own file
